refactor(curves): move curves to /utils
