+++ src/index.ts
@@ -8,8 +8,8 @@ export * from './lib/themes/theme';
 export { LIGHT_THEME } from './lib/themes/light_theme';
 export { DARK_THEME } from './lib/themes/dark_theme';
 export * from './lib/themes/theme_commons';
-export { RecursivePartial } from './lib/utils/commons';
-export { CurveType } from './lib/series/curves';
+export { RecursivePartial } from './utils/commons';
+export { CurveType } from './utils/curves';
 export { timeFormatter, niceTimeFormatter, niceTimeFormatByDay } from './utils/data/formatters';
 export { DataGenerator } from './utils/data_generators/data_generator';
 export { DataSeriesColorsValues } from './lib/series/series';
