+++ src/lib/series/specs.ts
@@ -10,7 +10,7 @@ import { Accessor } from '../utils/accessor';
 import { Omit, RecursivePartial } from '../../utils/commons';
 import { AnnotationId, AxisId, GroupId, SpecId } from '../../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../utils/scales/scales';
-import { CurveType } from './curves';
+import { CurveType } from '../../utils/curves';
 import { DataSeriesColorsValues } from './series';
 
 export type Datum = any;
